Use type float32 instead of float64 for regularization weights.

+++ cpp/subprojects/boosting/include/mlrl/boosting/rule_evaluation/rule_evaluation_decomposable_complete_binned.hpp
@@ -17,9 +17,9 @@ namespace boosting {
     class DecomposableCompleteBinnedRuleEvaluationFactory final : public IDecomposableRuleEvaluationFactory {
         private:
 
-            const float64 l1RegularizationWeight_;
+            const float32 l1RegularizationWeight_;
 
-            const float64 l2RegularizationWeight_;
+            const float32 l2RegularizationWeight_;
 
             const std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr_;
 
@@ -34,7 +34,7 @@ namespace boosting {
              *                                  allows to create the implementation to be used to assign labels to bins
              */
             DecomposableCompleteBinnedRuleEvaluationFactory(
-              float64 l1RegularizationWeight, float64 l2RegularizationWeight,
+              float32 l1RegularizationWeight, float32 l2RegularizationWeight,
               std::unique_ptr<ILabelBinningFactory> labelBinningFactoryPtr);
 
             std::unique_ptr<IRuleEvaluation<DenseDecomposableStatisticVector<float64>>> create(
