+++ cpp/subprojects/boosting/src/mlrl/boosting/rule_evaluation/rule_evaluation_non_decomposable_complete.cpp
@@ -5,7 +5,7 @@
 namespace boosting {
 
     NonDecomposableCompleteRuleEvaluationFactory::NonDecomposableCompleteRuleEvaluationFactory(
-      float64 l1RegularizationWeight, float64 l2RegularizationWeight, const Blas& blas, const Lapack& lapack)
+      float32 l1RegularizationWeight, float32 l2RegularizationWeight, const Blas& blas, const Lapack& lapack)
         : l1RegularizationWeight_(l1RegularizationWeight), l2RegularizationWeight_(l2RegularizationWeight), blas_(blas),
           lapack_(lapack) {}
